requirements.txt
Flask
line-bot-sdk
python-dotenv
google-generativeai
google-api-core==2.24.0
gunicorn
sqlalchemy
pymysql